move classes for sync to new 'sync' package
