+++ opencorpora/src/main/java/org/opencorpora/sync/TaskProvider.java
@@ -1,4 +1,4 @@
-package org.opencorpora;
+package org.opencorpora.sync;
 
 import android.content.ContentProvider;
 import android.content.ContentValues;
